Initial spinner with track artwork
Add a custom album spinner adaptor that can display images along
with album name.

Add a LRU cache with background image generation to populate the
spinner and change the playing info image to use the same LRU.
Media session track information is not using the new LRU as it
runs in a different task instance.

Signed-off-by: Tod Fitch <Tod@FitchDesign.com>